treat paths as commodities; test on Sioux Falls
